% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdistance_RPS.R
\name{cmdistance_RPS}
\alias{cmdistance_RPS}
\title{This function computes the least-squares Procrustes distance between
each pair of matrices (configurations of landmarks) from the input set}
\usage{
cmdistance_RPS(X)
}
\arguments{
\item{X}{The input set of nx3 matrices (objects)}
}
\value{
The LS Procrustes distance matrix between pairs of objects
}
\description{
This function computes the least-squares Procrustes distance between
each pair of matrices (configurations of landmarks) from the input set
}
\author{
Guillermo Pacheco, Viviana Ferraggine, Sebastian Torcida
}
