% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIntersections.R
\name{computeIntersections}
\alias{computeIntersections}
\title{Description}
\usage{
computeIntersections(X, ii, D)
}
\value{
s
}
\description{
Description
}
\examples{

}
\author{
Guillermo Andres Pacheco
}
