% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteLandmark.R
\name{deletePoint}
\alias{deletePoint}
\title{This procedure deletes a landmark (row) from the input nx3 matrix}
\usage{
deletePoint(x, num_landmark)
}
\arguments{
\item{x}{A nx3 matrix (n is the number of landmarks)}

\item{num_landmark}{The number of landmark (row) to be deleted}
}
\value{
x The corresponding (n-1)x3 output matrix
}
\description{
This procedure deletes a landmark (row) from the input nx3 matrix
}
\author{
Guillermo A. Pacheco, Viviana Ferraggine, Sebastian Torcida
}
