% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanciaR.R
\name{distanciaR}
\alias{distanciaR}
\title{This function computes the a resistant distance between
each pair of matrices from the input set}
\usage{
distanciaR(X)
}
\arguments{
\item{X}{The input set of nx3 matrices (objects)}
}
\value{
This function computes the a resistant distance between each pair of matrices from the input set
}
\description{
This function computes the a resistant distance between
each pair of matrices from the input set
}
\author{
Guillermo A. Pacheco, Viviana Ferraggine, Sebastian Torcida
}
