% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucunivMDS_RPS.R
\name{eucunivMDS_RPS}
\alias{eucunivMDS_RPS}
\title{Given a n x n distance matrix D (not necessarily Euclidean) and a initial
 set X0 of n seeds in k dim (that is, an initial n x k matrix), this function finds
 a set of n points in k dimensions X (a final n x k matrix) through a least-squares
 criterion such that the n x n matrix Dk of euclidean distances among these new points X
 is as close as possible to D.}
\usage{
eucunivMDS_RPS(D, k)
}
\arguments{
\item{D}{distance matrix n x n to be approximated}

\item{k}{dimension of output results}
}
\value{
X  A set of n points in k dimensions
}
\description{
Given a n x n distance matrix D (not necessarily Euclidean) and a initial
 set X0 of n seeds in k dim (that is, an initial n x k matrix), this function finds
 a set of n points in k dimensions X (a final n x k matrix) through a least-squares
 criterion such that the n x n matrix Dk of euclidean distances among these new points X
 is as close as possible to D.
}
\author{
Guillermo Pacheco, Viviana Ferraggine, Sebastian Torcida
}
