% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_symm.R
\name{matching.symm}
\alias{matching.symm}
\title{Description}
\usage{
matching.symm(A, ctr = "gmedian", side.file, legend.loc = "topleft")
}
\arguments{
\item{A}{s}

\item{ctr}{s}

\item{side.file}{s}

\item{legend.loc}{s}
}
\value{

}
\description{
Description
}
\author{
Guillermo A. Pacheco, Viviana Ferraggine, Federico Lotto, Sebastian Torcida
}
