% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_symm.R
\name{object.symm}
\alias{object.symm}
\title{Description}
\usage{
object.symm(A, ctr = "gmedian", prs.file, proj.met = "msum",
  legend.loc = "topleft")
}
\arguments{
\item{A}{s}

\item{ctr}{s}

\item{prs.file}{s}

\item{proj.met}{s}

\item{legend.loc}{s}
}
\value{

}
\description{
Description
}
\author{
Guillermo A. Pacheco, Viviana Ferraggine, Federico Lotto, Sebastian Torcida
}
