% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustesCM.R
\name{procrustesCM}
\alias{procrustesCM}
\title{This function performs the adjusting of the different specimens by least squares}
\usage{
procrustesCM(X)
}
\arguments{
\item{X}{A s-dimensional array of n * k matrix, representing the different specimens to be analyzed}
}
\value{
s-dimensional array of n * k matrix, representing the different specimens after adjusting.
}
\description{
This function performs the adjusting of the different specimens by least squares
}
\author{
Guillermo Andres Pacheco, Viviana Elizabeth Ferraggine, Sebastian Torcida
}
