% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustesCM_RPS.R
\name{procrustesCM_RPS}
\alias{procrustesCM_RPS}
\title{This function performs the classical least-squares (LS)
Procrustes superimposition of the input configurations of landmarks}
\usage{
procrustesCM_RPS(X)
}
\arguments{
\item{X}{A s-dimensional array (s=2 or s=3) of n x k matrices, representing shapes of k objects through n landmarks in s dimensions}
}
\value{
s-dimensional array of n x k matrices, representing the different specimens after adjusting.
}
\description{
This function performs the classical least-squares (LS)
Procrustes superimposition of the input configurations of landmarks
}
\author{
Guillermo Pacheco, Viviana Ferraggine, Sebastian Torcida
}
