% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeDistance.R
\name{relativeDistance}
\alias{relativeDistance}
\title{Description}
\usage{
relativeDistance(X, specm1, specm2, t)
}
\value{
s
}
\description{
Description
}
\examples{

}
\author{
Guillermo Andres Pacheco
}
