% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resdistance_RPS.R
\name{resdistance_RPS}
\alias{resdistance_RPS}
\title{This function computes the a resistant distance between
each pair of matrices from the input set}
\usage{
resdistance_RPS(X)
}
\arguments{
\item{X}{The input set of nx3 matrices (objects)}
}
\value{
This function computes the sum of non-squared euclidean distances across landmarks for each pair of matrices from the input set
}
\description{
This function computes the a resistant distance between
each pair of matrices from the input set
}
\author{
Guillermo A. Pacheco, Viviana Ferraggine, Sebastian Torcida
}
