% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resunivMDS_RPS.R
\name{resunivMDS_RPS}
\alias{resunivMDS_RPS}
\title{Given a n x n distance matrix D (not necessarily Euclidean) and a initial
 set X0 (that is, a n x k matrix) of n seeds in k dim, this function finds a set of
 n points in k dimensions X (that is, a k x n matrix) using a resistant criterion
 such that the n x n matrix Dk of euclidean distances among these new points X
 is as close as possible to D.}
\usage{
resunivMDS_RPS(D, k)
}
\arguments{
\item{D}{distance matrix n x n to be approximated}

\item{k}{dimension of output results}
}
\value{
X  A set of n points in k dimensions
}
\description{
Given a n x n distance matrix D (not necessarily Euclidean) and a initial
 set X0 (that is, a n x k matrix) of n seeds in k dim, this function finds a set of
 n points in k dimensions X (that is, a k x n matrix) using a resistant criterion
 such that the n x n matrix Dk of euclidean distances among these new points X
 is as close as possible to D.
}
\author{
Guillermo Pacheco, Viviana Ferraggine, Sebastian Torcida
 Given a n x n distance matrix D (not necessarily Euclidean) and a initial set X0 (nxk matrix)
 of n seeds in k dim, the function finds a set of n points  in k dimensions X (kxn matrix) by resistent method such
 that Euclidean distance nxn matrix Dk among these new points X is as close as to D.
}
