% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robgit.R
\docType{package}
\name{'Robgit'}
\alias{'Robgit'}
\alias{robgit}
\alias{'Robgit'-package}
\title{A generalized resistant Procrustes superposition}
\usage{
robgit(X, consenso = FALSE)
}
\arguments{
\item{X}{A s-dimensional array of n * k matrix, representing the different specimens to be analyzed}

\item{consenso}{A logical value that determines if the array consensus is returned.}
}
\value{
s-dimensional array of n * k matrix, representing the different specimens after adjusting.
}
\description{
A generalized resistant Procrustes superposition
}
\section{Functions}{

 deleteLandmark, univMDSrobust, univMDSeucl, distanciaCM, distanciaR, plotAllSpecimensRobust, readland.txtJ
}

\examples{
craniums <- readland.tps('DataSet/craneos2D.tps')
adjustment_craniums <- robgit2012(craniums, consenso = TRUE)
}
\author{
Guillermo Andres Pacheco
}
