% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robgit.R
\docType{package}
\name{'RPS'}
\alias{'RPS'}
\alias{robgit}
\alias{'RPS'-package}
\title{Resistant Procrustes Superposition in R (RPS_R),
a novel package for resistant morphometrics}
\usage{
robgit(X, consenso = FALSE)
}
\arguments{
\item{X}{A s-dimensional array of n * k matrix, representing the different specimens to be analyzed}

\item{consenso}{A logical value that determines if the array consensus is returned.}
}
\value{
s-dimensional array of n * k matrix, representing the different specimens after adjusting.
}
\description{
Description
Morphometric studies use landmark coordinates to
extract shape information, and shape differences between individuals are
analyzed by performing a Procrustes superposition of the corresponding
configurations of landmarks.
Which superposition criterion should be used is always a central issue;
typically, the least squares Procrustes superposition is chosen. This method minimizes
the sum of squared distances across landmarks, but it is widely accepted that the result can be
misleading whenever shape differences are located in less than 50% of the
landmarks.
A resistant Procrustes superposition (Torcida et al. 2014) is probably the most
elegant and efficient alternative, perfectly superimposing those landmarks
exhibiting no variation in case they are more than 50%. Despite this appealing
and rather intuitive feature, the resistant approach is seldom used probably due
to the lack of a free & friendly implementation.
This R-package specifically implements an a set of descriptive tools to perform
a resistant shape analysis of 2D and 3D configurations of landmarks.
The RPS_R package includes:
 *a generalized resistant Procrustes superposition (robgit.R)
 *a resistant distance to quantify the resulting shape differences, and
 *a resistant Multidimensional Scaling to obtain the corresponding ordination.
The corresponding least-squares counterparts have also been implemented, and both approaches can be therefore applied to the same dataset in order to compare their results.
}
\section{Functions}{

univMDSrobust, univMDSeucl, distanciaCM, distanciaR, readland.txtJ, robgit, procrustesCM
}

\author{
Guillermo Andres Pacheco, Viviana Elizabeth Ferraggine, Sebastian Torcida
}
